# packages/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: StreamYolo API
  version: 0.0.1
  description: Room-first live creator platform — rooms, tips, private sessions, tokens, CCBill.
servers:
  - url: http://localhost:3001
    description: Local dev
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Session token — delivered via Set-Cookie on login; also accepted as Bearer header for native clients.

  # ─── Reusable parameters ───────────────────────────────────────────────────
  parameters:
    Cursor:
      in: query
      name: cursor
      schema: { type: string }
      description: Opaque pagination cursor from previous page.
    Limit:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Insufficient role or ownership
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    PaginatedMeta:
      type: object
      required: [hasMore, nextCursor]
      properties:
        hasMore: { type: boolean }
        nextCursor:
          type: string
          nullable: true
...
    RoomCategory:
      type: string
      enum: [MALE, FEMALE, COUPLES, TRANS]
...
    PlatformSettingsDto:
      type: object
      required: [activePaymentProvider, tokenPurchasesEnabled]
      properties:
        activePaymentProvider: { type: string, enum: [CCBILL, DEMO] }
        tokenPurchasesEnabled: { type: boolean }
...
paths:
...
  /rooms:
    get:
      operationId: listRooms
      summary: List live public rooms
      security: []
      tags: [rooms]
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [LIVE, ENDED] }
        - in: query
          name: category
          schema:
            type: array
            items: { $ref: '#/components/schemas/RoomCategory' }
          style: form
          explode: true
        - in: query
          name: country
          schema:
            type: array
            items: { type: string, minLength: 2, maxLength: 2 }
          style: form
          explode: true
        - in: query
          name: tag
          schema:
            type: array
            items: { type: string }
          style: form
          explode: true
        - in: query
          name: q
          schema: { type: string }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated room list
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/RoomCard' }
                  meta: { $ref: '#/components/schemas/PaginatedMeta' }
...
  /creator/rooms/{roomId}/end:
    post:
      operationId: endRoom
      summary: End the creator's live room
      tags: [creator-rooms]
      parameters:
        - in: path
          name: roomId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Room ended
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [room]
                    properties:
                      room: { $ref: '#/components/schemas/RoomDetail' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Creator profile ──────────────────────────────────────────────────────────
...
  /rooms/{roomId}:
    get:
      operationId: getRoom
      summary: Get a public room by ID
      security: []
      tags: [rooms]
      parameters:
        - in: path
          name: roomId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Room detail with VIP user list
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [room]
                    properties:
                      room: { $ref: '#/components/schemas/RoomDetail' }
                      vipUserIds:
                        type: array
                        items: { type: string }
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Chat ─────────────────────────────────────────────────────────────────────
...
  /wallet:
    get:
      operationId: getWallet
      summary: Get the authenticated user's wallet and recent ledger entries
      tags: [wallet]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Wallet and ledger
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [wallet, ledger, meta]
                    properties:
                      wallet: { $ref: '#/components/schemas/WalletSummary' }
                      ledger:
                        type: array
                        items: { $ref: '#/components/schemas/LedgerEntryDto' }
                      meta: { $ref: '#/components/schemas/PaginatedMeta' }
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── Private sessions ──────────────────────────────────────────────────────────
...
  /media/{mediaId}:
    get:
      operationId: getMedia
      summary: Get media redirect
      tags: [media]
      security: []
      parameters:
        - in: path
          name: mediaId
          required: true
          schema: { type: string }
      responses:
        '302':
          description: Redirect
        '404':
          $ref: '#/components/responses/NotFound'
...
  /admin/overview:
    get:
      operationId: getAdminOverview
      summary: Admin dashboard overview — live rooms, user counts, pending queues
      tags: [admin]
      responses:
        '200':
          description: Admin overview metrics
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/AdminOverviewMetrics' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/settings:
    get:
      operationId: getAdminSettings
      summary: Get platform settings (admin)
      tags: [admin]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlatformSettingsDto' }
    patch:
      operationId: updateAdminSettings
      summary: Update platform settings (admin)
      tags: [admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PlatformSettingsDto' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlatformSettingsDto' }
  /admin/rooms:
    get:
      operationId: listAdminRooms
      summary: Admin list all rooms
      tags: [admin]
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [DRAFT, LIVE, ENDED, SUSPENDED, HIDDEN] }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Rooms list
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/RoomCard' }
                  meta: { $ref: '#/components/schemas/PaginatedMeta' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/rooms/{roomId}:
    get:
      operationId: getAdminRoom
      summary: Admin get room detail
      tags: [admin]
      parameters:
        - in: path
          name: roomId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Room detail
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/RoomDetail' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/rooms/{roomId}/end:
    post:
      operationId: adminEndRoom
      summary: Admin force-end a live room
      tags: [admin]
      parameters:
        - in: path
          name: roomId
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EndRoomInput' }
      responses:
        '200':
          description: Room ended
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [room]
                    properties:
                      room: { $ref: '#/components/schemas/RoomDetail' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/rooms/{roomId}/hide:
    post:
      operationId: adminHideRoom
      summary: Admin hide a room from public listing
      tags: [admin]
      parameters:
        - in: path
          name: roomId
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EndRoomInput' }
      responses:
        '200':
          description: Room hidden
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [room]
                    properties:
                      room: { $ref: '#/components/schemas/RoomDetail' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/users:
    get:
      operationId: listAdminUsers
      summary: Admin list all users
      tags: [admin]
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [ACTIVE, SUSPENDED, DELETED] }
        - in: query
          name: role
          schema: { type: string, enum: [VIEWER, CREATOR, ADMIN, MODERATOR] }
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Users list
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/PublicUser' }
                  meta: { $ref: '#/components/schemas/PaginatedMeta' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/users/{userId}:
    get:
      operationId: getAdminUser
      summary: Admin get user detail with wallet and creator profile
      tags: [admin]
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: User detail
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/AdminUserDetail' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/users/{userId}/wallet:
    get:
      operationId: getAdminUserWallet
      summary: Get user wallet and ledger
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Wallet and ledger list
          content:
            application/json:
              schema:
                type: object
                required: [wallet, ledger, meta]
                properties:
                  wallet: { $ref: '#/components/schemas/WalletSummary' }
                  ledger:
                    type: array
                    items: { $ref: '#/components/schemas/LedgerEntryDto' }
                  meta: { $ref: '#/components/schemas/PaginatedMeta' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/users/{userId}/wallet/grant:
    post:
      operationId: adminGrantTokens
      summary: Grant test tokens
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amountTokens, reason]
              properties:
                amountTokens: { type: integer, minimum: 1 }
                reason: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WalletSummary' }
        '400': { $ref: '#/components/responses/UnprocessableEntity' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/users/{userId}/wallet/revoke:
    post:
      operationId: adminRevokeTokens
      summary: Revoke test tokens
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amountTokens, reason]
              properties:
                amountTokens: { type: integer, minimum: 1 }
                reason: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WalletSummary' }
        '400': { $ref: '#/components/responses/UnprocessableEntity' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/users/{userId}/wallet/reset:
    post:
      operationId: adminResetWallet
      summary: Reset test tokens to 0
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WalletSummary' }
        '400': { $ref: '#/components/responses/UnprocessableEntity' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...